+++ scripts/clear-jit-cache.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+rm $(find . -type d -name __pashmam__) -rf
+rm $(find . -type d -name __pycache__) -rf