+++ src/core/program.py
@@ -31,7 +31,7 @@ from . import helpers, version, modules, jit, parser, current_prog
 from .class_system import Class, ClassObject
 from .function import Function
 
-import hashlib, time, random, datetime, json
+import hashlib, time, random, datetime, base64, json, http, http.cookies, http.server, http.client, http.cookiejar
 
 class Program(helpers.Helpers):
     """ Pashmak program object """
@@ -72,6 +72,11 @@ class Program(helpers.Helpers):
 
         self.out_started = False
         self.out_content = ''
+
+        self.current_func = []
+        self.current_class = []
+
+        self.func_depth = 0
 
         current_prog.current_prog = self
 
@@ -272,40 +277,7 @@ class Program(helpers.Helpers):
 
     def eval(self, command, only_parse=False, varname_as_dict=False, only_str_parse=False, dont_check_vars=False):
         """ Runs eval on command """
-        i = 0
-        command = command.strip()
-        is_in_string = False
-        command_parts = [[False, '']]
-        while i < len(command):
-            is_string_start = False
-            if command[i] == '"' or command[i] == "'":
-                before_backslashes_count = 0
-                try:
-                    x = i-1
-                    while x >= 0:
-                        if command[x] == '\\':
-                            before_backslashes_count += 1
-                        else:
-                            x = -1
-                        x -= 1
-                except:
-                    pass
-                if is_in_string:
-                    if before_backslashes_count % 2 != 0 and before_backslashes_count != 0:
-                        pass
-                    elif is_in_string == command[i]:
-                        is_in_string = False
-                        command_parts[-1][1] += command[i]
-                        is_string_start = True
-                        command_parts.append([False, ''])
-                else:
-                    is_in_string = command[i]
-                    command_parts.append([True, ''])
-                    command_parts[-1][1] += command[i]
-                    is_string_start = True
-            if not is_string_start:
-                command_parts[-1][1] += command[i]
-            i += 1
+        command_parts = parser.parse_string(command)
 
         if only_str_parse:
             return command_parts
@@ -374,29 +346,27 @@ class Program(helpers.Helpers):
         op = self.set_command_index(op)
         op_name = op['command']
 
+        if op_name == 'func':
+            self.func_depth += 1
+
         if op_name == 'endfunc':
-            self.run_endfunc(op)
+            if self.func_depth <= 1:
+                self.run_endfunc(op)
+                self.func_depth -= 1
+                return
+            self.func_depth -= 1
+
+        # if a function is started, append current command to the function body
+        if len(self.current_func) > 0:
+            if len(self.current_class) > 0:
+                self.classes[self.current_class[-1]].__methods__[self.current_func[-1]].body.append(op)
+            else:
+                self.functions[self.current_func[-1]].body.append(op)
             return
 
         if op_name == 'endclass':
             self.run_endclass(op)
             return
-
-        # if a function is started, append current command to the function body
-        try:
-            self.current_func
-            try:
-                self.current_class
-                self.classes[self.current_class].__methods__[self.current_func].body.append(op)
-            except:
-                self.functions[self.current_func].body.append(op)
-            return
-        except NameError:
-            pass
-        except KeyError:
-            pass
-        except AttributeError:
-            pass
 
         # list of commands
         commands_dict = {
@@ -436,15 +406,8 @@ class Program(helpers.Helpers):
         if op['str'][0] == '$':
             # if a class is started, append current command as a property to class
             is_in_class = False
-            try:
-                self.current_class
+            if len(self.current_class) > 0:
                 is_in_class = True
-            except NameError:
-                pass
-            except KeyError:
-                pass
-            except AttributeError:
-                pass
             parts = self.split_by_equals(op['str'].strip())
             if len(parts) <= 1:
                 if '->' in op['str'] or '(' in op['str'] or ')' in op['str']:
@@ -459,7 +422,7 @@ class Program(helpers.Helpers):
             varname = varname[0]
             if len(parts) <= 1:
                 if is_in_class:
-                    self.classes[self.current_class].__props__[varname[1:]] = None
+                    self.classes[self.current_class[-1]].__props__[varname[1:]] = None
                 else:
                     self.set_var(varname[1:], None)
                 return
@@ -471,7 +434,7 @@ class Program(helpers.Helpers):
                 exec('tmp_real_var.__props__.' + is_class_setting + ' = value')
             else:
                 if is_in_class:
-                    self.classes[self.current_class].__props__[varname[1:]] = value
+                    self.classes[self.current_class[-1]].__props__[varname[1:]] = value
                 else:
                     if '[' in varname or ']' in varname:
                         the_target = self.eval(varname, only_parse=True)
@@ -490,7 +453,8 @@ class Program(helpers.Helpers):
         # check function exists
         func_real_name = self.get_func_real_name(op_name)
         if func_real_name == False:
-            return self.raise_error('SyntaxError', 'undefined function "' + op_name + '"', op)
+            self.mem = self.eval(op['str'])
+            return
         func_body = self.functions[func_real_name]
 
         # run function